+++ Commands/SetShooterHeightCommand.h
@@ -5,12 +5,12 @@
 class SetShooterHeightCommand : public DefaultCommand {
 
 public:
-	SetShooterHeightCommand(int height);
+	SetShooterHeightCommand(double height);
 	void Initialize() override;
 	bool IsFinished() override;
 	void End() override;
 
 protected:
-	int m_height;
+	double m_height;
 	int m_speed;
 };
